StarsFix
A flurry of changes related to the ShareShip.pl (and associated functions in TH). 

Fixing Stars! files is complicated, because each time a user makes a change to a .x file, it's just added to the list of all the other changes, and sorted out at the Host. This now correctly detects not only bug entries, but the correction of bug entries, so you don't have to delete the .x file to get the errors to stop showing up. It was a real pain. But it's cool now! and I have a model for future fixes.